Refactor lanelet::time utility functionalities to be public (#223)
* refactor

* add namespace to carma-traffic-signal

* typo

* typo

* header change
